+++ CHANGELOG.md
@@ -13,6 +13,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   `Hash`.
 - `Entry`, the type which `Object` is a collection of, is now exported. This was
   an oversight. Thanks to @PolScript for reporting this in #6.
+- `Value::get(key)`, `Value::get_mut(key)`, and `Index<&str>`/`IndexMut<&str>`
+  have been added to allow easier access to the contents of a `Value` when it is
+  an object. Similar to most built-in collection types, the `Index` trait
+  implementations will panic if the key is not found, but the `get` variants
+  return an Option instead of panicking.
+- `Value::get_index(usize)`, `Value::get_index_mut(usize)`, and
+  `Index<usize>`/`IndexMut<usize>` have been added to allow easier access to the
+  contents of a `Value` when it is an array. Similar to most built-in collection
+  types, the `Index` trait implementations will panic if the key is not found,
+  but the `get` variants return an Option instead of panicking.
 
 ## v0.2.1